exit with error message on failure to mmap V4L device
Whem mmap() fails it returns -1 (0xfffff...).
This later becomes a SEGFAULT when trying to memcpy this address.
Detect the failure early on an exit with (hopefully) helpful message.

This can happen if the LD_PRELOAD pointing to the V4L shared object
is required but missing (can happen e.g. when loading with empty environment
under systemd).w